Clear extension settings when snooze ends
Using the Clear() api call instead of creating a new rule to Allow All should effectively allow users to toggle between Ask and Block All modes, which is what I've strived for all along. I don't know if clear() was always available, or if it's a new addition since I last looked at the api, but I'm glad it seems to work for my purposes.